home work soring data

+++ 2022-08/jpa-home-work-new/src/main/java/ru/otus/spring/service/GenreService.java
@@ -12,12 +12,12 @@ public interface GenreService {
 
     Genre insert(String genre);
 
-    Genre updateById(String genre, int id);
+    Genre updateById(String genre, long id);
 
-    void deleteById(int id);
+    void deleteById(long id);
 
     List<Genre> getAll();
 
-    Optional<Genre> getById(int id);
+    Optional<Genre> getById(long id);
 
 }
